feat: add virtual storage cell
